requirements.txt
PyGithub==2.5.0
python-dotenv==1.0.1
pandas==2.2.3
scikit-learn==1.6.0
streamlit==1.41.1
tqdm==4.67.1
matplotlib==3.10.0
seaborn==0.13.2
requests==2.32.3
xgboost==1.7.6